Split entity/filter and use entity in people
